LPS-61652 migrate to internal package
