+++ modules/portal-security/portal-security-sso-ntlm/src/main/java/com/liferay/portal/security/sso/ntlm/internal/Netlogon.java
@@ -12,15 +12,18 @@
  * details.
  */
 
-package com.liferay.portal.security.sso.ntlm;
+package com.liferay.portal.security.sso.ntlm.internal;
 
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.security.sso.ntlm.msrpc.NetlogonAuthenticator;
-import com.liferay.portal.security.sso.ntlm.msrpc.NetlogonIdentityInfo;
-import com.liferay.portal.security.sso.ntlm.msrpc.NetlogonNetworkInfo;
-import com.liferay.portal.security.sso.ntlm.msrpc.NetlogonValidationSamInfo;
-import com.liferay.portal.security.sso.ntlm.msrpc.NetrLogonSamLogon;
+import com.liferay.portal.security.sso.ntlm.NetlogonConnectionManager;
+import com.liferay.portal.security.sso.ntlm.NtlmLogonException;
+import com.liferay.portal.security.sso.ntlm.NtlmServiceAccount;
+import com.liferay.portal.security.sso.ntlm.internal.msrpc.NetlogonAuthenticator;
+import com.liferay.portal.security.sso.ntlm.internal.msrpc.NetlogonIdentityInfo;
+import com.liferay.portal.security.sso.ntlm.internal.msrpc.NetlogonNetworkInfo;
+import com.liferay.portal.security.sso.ntlm.internal.msrpc.NetlogonValidationSamInfo;
+import com.liferay.portal.security.sso.ntlm.internal.msrpc.NetrLogonSamLogon;
 
 import java.io.IOException;
