+++ modules/portal-security/portal-security-sso-ntlm/src/main/java/com/liferay/portal/security/sso/ntlm/internal/NtlmManager.java
@@ -12,10 +12,13 @@
  * details.
  */
 
-package com.liferay.portal.security.sso.ntlm;
+package com.liferay.portal.security.sso.ntlm.internal;
 
 import com.liferay.portal.kernel.util.ArrayUtil;
 import com.liferay.portal.kernel.util.StringPool;
+import com.liferay.portal.security.sso.ntlm.NetlogonConnectionManager;
+import com.liferay.portal.security.sso.ntlm.NtlmLogonException;
+import com.liferay.portal.security.sso.ntlm.NtlmServiceAccount;
 
 import java.io.IOException;
 import java.io.UnsupportedEncodingException;
