+++ modules/portal-security/portal-security-sso-ntlm/src/main/java/com/liferay/portal/security/sso/ntlm/internal/module/configuration/definition/NtlmCompanyServiceConfigurationPidMapping.java
@@ -12,10 +12,11 @@
  * details.
  */
 
-package com.liferay.portal.security.sso.ntlm.configuration.definition;
+package com.liferay.portal.security.sso.ntlm.internal.module.configuration.definition;
 
-import com.liferay.portal.kernel.settings.definition.ConfigurationBeanDeclaration;
+import com.liferay.portal.kernel.settings.definition.ConfigurationPidMapping;
 import com.liferay.portal.security.sso.ntlm.configuration.NtlmConfiguration;
+import com.liferay.portal.security.sso.ntlm.constants.NtlmConstants;
 
 import org.osgi.service.component.annotations.Component;
 
@@ -23,12 +24,17 @@ import org.osgi.service.component.annotations.Component;
  * @author Mika Koivisto
  */
 @Component
-public class NtlmCompanyServiceConfigurationBeanDeclaration
-	implements ConfigurationBeanDeclaration {
+public class NtlmCompanyServiceConfigurationPidMapping
+	implements ConfigurationPidMapping {
 
 	@Override
 	public Class<?> getConfigurationBeanClass() {
 		return NtlmConfiguration.class;
+	}
+
+	@Override
+	public String getConfigurationPid() {
+		return NtlmConstants.SERVICE_NAME;
 	}
 
 }
