+++ modules/portal-security/portal-security-sso-ntlm/src/main/java/com/liferay/portal/security/sso/ntlm/internal/msrpc/NetlogonAuthenticator.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.security.sso.ntlm.msrpc;
+package com.liferay.portal.security.sso.ntlm.internal.msrpc;
 
 import jcifs.dcerpc.ndr.NdrBuffer;
 import jcifs.dcerpc.ndr.NdrObject;
@@ -20,41 +20,52 @@ import jcifs.dcerpc.ndr.NdrObject;
 /**
  * @author Marcellus Tavares
  */
-public class GroupMembership extends NdrObject {
+public class NetlogonAuthenticator extends NdrObject {
 
-	public GroupMembership() {
+	public NetlogonAuthenticator() {
+		_credential = new byte[8];
 	}
 
-	public GroupMembership(int relativeId, int attributes) {
-		_relativeId = relativeId;
-		_attributes = attributes;
+	public NetlogonAuthenticator(byte[] credential, int timestamp) {
+		_credential = credential;
+		_timestamp = timestamp;
 	}
 
 	@Override
 	public void decode(NdrBuffer ndrBuffer) {
 		ndrBuffer.align(4);
 
-		_relativeId = ndrBuffer.dec_ndr_long();
-		_attributes = ndrBuffer.dec_ndr_long();
+		int index = ndrBuffer.index;
+
+		ndrBuffer.advance(8);
+
+		_timestamp = ndrBuffer.dec_ndr_long();
+
+		ndrBuffer = ndrBuffer.derive(index);
+
+		for (int i = 0; i < 8; i++) {
+			_credential[i] = (byte)ndrBuffer.dec_ndr_small();
+		}
 	}
 
 	@Override
 	public void encode(NdrBuffer ndrBuffer) {
 		ndrBuffer.align(4);
 
-		ndrBuffer.enc_ndr_long(_relativeId);
-		ndrBuffer.enc_ndr_long(_attributes);
+		int index = ndrBuffer.index;
+
+		ndrBuffer.advance(8);
+
+		ndrBuffer.enc_ndr_long(_timestamp);
+
+		ndrBuffer = ndrBuffer.derive(index);
+
+		for (int i = 0; i < 8; i++) {
+			ndrBuffer.enc_ndr_small(_credential[i]);
+		}
 	}
 
-	public int getAttributes() {
-		return _attributes;
-	}
-
-	public int getRelativeId() {
-		return _relativeId;
-	}
-
-	private int _attributes;
-	private int _relativeId;
+	private final byte[] _credential;
+	private int _timestamp;
 
 }
