+++ modules/portal-security/portal-security-sso-ntlm/src/main/java/com/liferay/portal/security/sso/ntlm/internal/msrpc/NetrServerReqChallenge.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.security.sso.ntlm.msrpc;
+package com.liferay.portal.security.sso.ntlm.internal.msrpc;
 
 import jcifs.dcerpc.DcerpcMessage;
 import jcifs.dcerpc.ndr.NdrBuffer;
@@ -20,20 +20,16 @@ import jcifs.dcerpc.ndr.NdrBuffer;
 /**
  * @author Marcellus Tavares
  */
-public class NetrServerAuthenticate3 extends DcerpcMessage {
+public class NetrServerReqChallenge extends DcerpcMessage {
 
-	public NetrServerAuthenticate3(
-		String primaryName, String accountName, int secureChannelType,
-		String computerName, byte[] clientCredential, byte[] serverCredential,
-		int negotiateFlags) {
+	public NetrServerReqChallenge(
+		String primaryName, String computerName, byte[] clientChallenge,
+		byte[] serverChallenge) {
 
 		_primaryName = primaryName;
-		_accountName = accountName;
-		_secureChannelType = (short)secureChannelType;
 		_computerName = computerName;
-		_clientCredential = clientCredential;
-		_serverCredential = serverCredential;
-		_negotiateFlags = negotiateFlags;
+		_clientChallenge = clientChallenge;
+		_serverChallenge = serverChallenge;
 
 		ptype = 0;
 		flags = DCERPC_FIRST_FRAG | DCERPC_LAST_FRAG;
@@ -48,11 +44,9 @@ public class NetrServerAuthenticate3 extends DcerpcMessage {
 		ndrBuffer = ndrBuffer.derive(index);
 
 		for (int i = 0; i < 8; i++) {
-			_serverCredential[i] = (byte)ndrBuffer.dec_ndr_small();
+			_serverChallenge[i] = (byte)ndrBuffer.dec_ndr_small();
 		}
 
-		_negotiateFlags = ndrBuffer.dec_ndr_long();
-		_accountRid = ndrBuffer.dec_ndr_long();
 		_status = ndrBuffer.dec_ndr_long();
 	}
 
@@ -60,52 +54,36 @@ public class NetrServerAuthenticate3 extends DcerpcMessage {
 	public void encode_in(NdrBuffer ndrBuffer) {
 		ndrBuffer.enc_ndr_referent(_primaryName, 1);
 		ndrBuffer.enc_ndr_string(_primaryName);
-		ndrBuffer.enc_ndr_string(_accountName);
-		ndrBuffer.enc_ndr_short(_secureChannelType);
 		ndrBuffer.enc_ndr_string(_computerName);
 
 		int index = ndrBuffer.index;
 
 		ndrBuffer.advance(8);
 
-		NdrBuffer derivedNrdBuffer = ndrBuffer.derive(index);
+		ndrBuffer = ndrBuffer.derive(index);
 
 		for (int i = 0; i < 8; i++) {
-			derivedNrdBuffer.enc_ndr_small(_clientCredential[i]);
+			ndrBuffer.enc_ndr_small(_clientChallenge[i]);
 		}
-
-		ndrBuffer.enc_ndr_long(_negotiateFlags);
-	}
-
-	public int getAccountRid() {
-		return _accountRid;
-	}
-
-	public int getNegotiatedFlags() {
-		return _negotiateFlags;
 	}
 
 	@Override
 	public int getOpnum() {
-		return 26;
+		return 4;
 	}
 
-	public byte[] getServerCredential() {
-		return _serverCredential;
+	public byte[] getServerChallenge() {
+		return _serverChallenge;
 	}
 
 	public int getStatus() {
 		return _status;
 	}
 
-	private final String _accountName;
-	private int _accountRid;
-	private final byte[] _clientCredential;
+	private final byte[] _clientChallenge;
 	private final String _computerName;
-	private int _negotiateFlags;
 	private final String _primaryName;
-	private final short _secureChannelType;
-	private final byte[] _serverCredential;
+	private final byte[] _serverChallenge;
 	private int _status;
 
 }
